Visualizer is functioning
I ended up having to find a audio converter.  Also unfortunately the
audio converter I found that worked is part of another visualizer.
Essentially what I did was take out the visual part of his visualizer
and replaced it with a three.js version of spheres in a ring that spin
around and move in and out with their associated audio values.  As of
now I would like to add shading to the spheres to give it more feel, and
the scene started bogging down when I used spheres with 30 subdivisions
along either axis.  I lowered the subdivisions down to 10 and it runs
fine, but I don't know how that'll work with the time variable I wanted
to add.  Removed the audio file as now this runs on an uploaded audio
file (I know mp3's work, but haven't tried much else